also sort dictionary list by custom order
